+++ src/md/changelog.md
@@ -1,6 +1,8 @@
 # staging
 
-* Account password can now be changed in the settings page (finally…).
+* Account password can now be changed in the settings page
+  (finally…). Settings page now has tabbed sections for easier
+  browsing.
 
 # Version 0.10